requirements.txt
fastapi[standard]
pydantic
uvicorn
paho-mqtt